ft_memchr complete

+++ LibFt/ft_memset.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, int len)
+void	*ft_memset(void *s, int c, size_t len)
 {
 	char	*str;
